+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### PVE API client
 #### Added
 - `POST   /custom-api/v1/lxc/{id}/exec-async` implementation as `PVE.LXC.ExecAsync`.
+- `GET    /custom-api/v1/cmd/{id}` implementation as `PVE.LXC.GetCMDResult`.
 
 ## [v0.6.1]
 ### PVE API client